requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2021.1.19=haa95532_0
certifi=2020.12.5=py38haa95532_0
imutils=0.5.4=pypi_0
numpy=1.20.1=pypi_0
opencv-contrib-python=4.5.1.48=pypi_0
openssl=1.1.1j=h2bbff1b_0
pillow=8.1.1=pypi_0
pip=20.3.3=py38haa95532_0
python=3.8.5=h5fd99cc_1
pytz=2021.1=pypi_0
setuptools=52.0.0=py38haa95532_0
sqlite=3.33.0=h2a8f88b_0
vc=14.2=h21ff451_1
vptree=1.2=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
zlib=1.2.11=h62dcd97_4
